+++ collectors/conf/postgresql.conf
@@ -1,7 +1,10 @@
 [base]
 enabled = False
 interval = 30
-
-host =
+host = localhost
 user = cloudwiz_user
 password = cloudwiz_pass
+#dbname = testdb
+#relations = ['test']
+
+
